lang/hermes-runtime-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-runtime"
version = "0.1.0"
description = "Grid EAASP L1 Runtime — hermes-agent T2 Aligned adapter"
requires-python = ">=3.11"
dependencies = [
  "grpcio>=1.80.0,<2",
  "grpcio-tools>=1.80.0,<2",
  "protobuf>=6.31.1,<7",
  "python-dotenv>=1.0.0,<2",
]

[project.optional-dependencies]
dev = ["pytest>=9.0.0,<10", "pytest-asyncio>=1.0.0,<2"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
